Split `outcome_mapper` into 3 files
