+++ src/domain/resources/facade/src/facade/graphql/outcome_mapper/mod.rs
@@ -1,0 +1,15 @@
+// Copyright Kamu Data, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+mod batch_helpers;
+mod operation_outcomes;
+mod problem_mappers;
+
+pub(crate) use operation_outcomes::*;
+pub(crate) use problem_mappers::{bad_account_problem_error, unsupported_descriptor_problem_error};
